requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
amqp==1.4.9
ansi2html==1.4.2
anyjson==0.3.3
apipkg==1.5
appdirs==1.4.3
Appium-Python-Client==0.26
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.7.0
backports.csv==1.0.7
backports.functools-lru-cache==1.5
bcrypt==3.1.7
beautifulsoup4==4.6.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.26.post2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cheroot==6.5.4
CherryPy==18.1.0
click==6.7
colorama==0.3.9
colorlog==4.0.2
concurrent-log-handler==0.9.9
configparser==5.0.0
cryptography==2.6.1
cryptokit==0.0.7
Django==2.0.4
elasticsearch==7.9.1
execnet==1.7.1
Flask==1.1.2
har2case==0.3.1
HttpRunner==1.5.8
idna==2.8
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
jaraco.functools==2.0
Jinja2==2.11.2
jsonpath==0.82
kafka-python==1.4.6
kombu==3.0.37
MarkupSafe==1.1.1
more-itertools==6.0.0
nb-log==3.4
packaging==20.4
pika==1.1.0
pluggy==0.13.1
portend==2.3
py==1.9.0
pycparser==2.20
pycryptodome==3.9.8
pymongo==3.5.1
PyMySQL==0.9.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
pytest==6.0.2
pytest-base-url==1.4.2
pytest-forked==1.3.0
pytest-html==1.19.0
pytest-metadata==1.10.0
pytest-repeat==0.7.0
pytest-rerunfailures==9.1
pytest-xdist==1.23.2
pytz==2018.9
pywin32==224
PyYAML==3.12
requests==2.21.0
requests-toolbelt==0.9.1
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.3
selenium==2.53.6
six==1.12.0
tempora==1.14
toml==0.10.1
tomorrow3==1.1.0
urllib3==1.24.1
wcwidth==0.2.5
Werkzeug==1.0.1
yaml-1.3==0.1.0
zc.lockfile==1.4
zipp==3.1.0
